Better circle size scale, rotated axis, tooltip almost done
Still need to make the svg and the buttons stay in the middle of the screen when scrolling and zooming in/out with the browser. Need to align half of the tooltip text to the left, one half is already aligned to the right. Need to make the image better/bigger